pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "cffi==2.0.0",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.1",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.117.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "idna==3.10",
    "passlib==1.7.4",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.11.0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "starlette==0.48.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "uvicorn==0.37.0",
]
